+++ actix-tls/tests/test_resolvers.rs
@@ -38,8 +38,9 @@ async fn custom_resolver() {
 async fn custom_resolver_connect() {
     use trust_dns_resolver::TokioAsyncResolver;
 
-    let srv =
-        TestServer::with(|| fn_service(|_io: TcpStream| async { Ok::<_, io::Error>(()) }));
+    let srv = TestServer::with(fn_service(|_io: TcpStream| async {
+        Ok::<_, io::Error>(())
+    }));
 
     struct MyResolver {
         trust_dns: TokioAsyncResolver,